Beta function removed and version bumped to 0.9.
The beta function is complex enough that I don't feel comfortable using the implementation here.  There are too many special cases it should handle but doesn't, and _can't_ with the current signature.  To handle them all correctly (especially if Complex arguments are to be supported), it really needs to be broken into multiple functions.  At that point, one might as well use a type class as in the under-construction 'beta' package.